pyproject.toml
[tool.poetry]
name = "freeseat-api"
version = "0.1.0"
description = "API for application to connect people who require a lift with ones who can give it."
authors = ["Alex Rozum <rozumalex@gmail.com>"]
license = "GNU General Public License v3.0"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.3"
psycopg2 = "^2.9.3"
gunicorn = "^20.1.0"
djangorestframework = "^3.13.1"
django-filter = "^21.1"
django-simple-history = "^3.0.0"
django-modeltranslation = "^0.17.5"
djangorestframework-gis = "^0.18"
Pillow = "^9.0.1"
drf-spectacular = "^0.21.2"
django-picklefield = "^3.0.1"
django-admin-geomap = {git = "https://github.com/julleks/django.admin.geomap.git", rev = "main"}
sentry-sdk = "^1.5.6"
django-cors-headers = "^3.11.0"
newrelic = "^7.6.0"
pytest-django = "^4.5.2"
model-bakery = "^1.4.0"
django-treebeard = "^4.5.1"

[tool.poetry.dev-dependencies]
pre-commit = "^2.17.0"
commitizen = "^2.21.2"
pytest = "^7.0.1"
pytest-cov = "^3.0.0"

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"
tag_format = "$version"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
